Fix bug when "up" is a "feed" in Terraform
If an answer's `meta` field's `up` value is not a bool but a feed,
it causes a panic and crashes terraform.

related to this issue:
https://github.com/terraform-providers/terraform-provider-ns1/issues/31

Add new case where interface could be a `map[string]interface{}`.

Also, allow the `up` value to be something other than a bool, since
it can be a feed.